Chapter 10, 11, 12

+++ Project/classes/Ijdb/Controllers/Joke.php
@@ -2,14 +2,17 @@
     namespace Ijdb\Controllers;
 
     use \Hanbit\DatabaseTable;
+    use \Hanbit\Authentication;
 
     class Joke {
         private $jokesTable;
         private $authorsTable;
+        private $authentication;
 
-        public function __construct(DatabaseTable $jokesTable, DatabaseTable $authorsTable) {
+        public function __construct(DatabaseTable $jokesTable, DatabaseTable $authorsTable, Authentication $authentication) {
             $this -> jokesTable = $jokesTable;
             $this -> authorsTable = $authorsTable;
+            $this -> authentication = $authentication;
         }
 
         public function home() {
@@ -34,7 +37,8 @@
                     'joketext' => $joke['joketext'],
                     'jokedate' => $joke['jokedate'],
                     'name' => $author['name'],
-                    'email' => $author['email']
+                    'email' => $author['email'],
+                    'authorId' => $author['id']
                 ];
             }
 
@@ -42,20 +46,32 @@
 
             $totalJokes = $this -> jokesTable -> total();
 
+            $author = $this -> authentication -> getUser();
+
             return [
                 'template' => 'jokes.html.php',
                 'title' => $title,
                 'variables' => [
                     'totalJokes' => $totalJokes,
-                    'jokes' => $jokes
+                    'jokes' => $jokes,
+                    'userId' => $author['id'] ?? null
                 ]
             ];
         }
 
         public function saveEdit() {
+            $author = $this -> authentication -> getUser();
+
+            if (isset($_GET['id'])) {
+                $joke = $this -> jokesTable -> findById($_GET['id']);
+                if ($joke['authorId'] != $author['id']) {
+                    return;
+                }
+            }
+
             $joke = $_POST['joke'];
             $joke['jokedate'] = new \DateTime();
-            $joke['authorId'] = 1;
+            $joke['authorId'] = $author['id'];
 
             $this -> jokesTable -> save($joke);
 
@@ -63,6 +79,8 @@
         }
 
         public function edit() {
+            $author = $this -> authentication -> getUser();
+
             $title = '유머 글 등록';
 
             if (isset($_GET['id'])) {
@@ -75,12 +93,21 @@
                 'template' => 'editjoke.html.php',
                 'title' => $title,
                 'variables' => [
-                    'joke' => $joke ?? null
+                    'joke' => $joke ?? null,
+                    'userId' => $author['id'] ?? null
                 ]
             ];
         }
 
         public function delete() {
+            $author = $this -> authentication -> getUser();
+
+            $joke = $this -> jokesTable -> findById($_POST['id']);
+
+            if ($joke['authorId'] != $author['id']) {
+                return;
+            }
+
             $this -> jokesTable -> delete($_POST['id']);
 
             header('location: /joke/list');
